Create storyboard
