+++ Reading List again/Reading List again/ReadingListViewController.swift
@@ -1,0 +1,30 @@
+//
+//  ReadingListViewController.swift
+//  Reading List again
+//
+//  Created by Nathanael Youngren on 1/17/19.
+//  Copyright © 2019 Nathanael Youngren. All rights reserved.
+//
+
+import UIKit
+
+class ReadingListViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
