+++ Reading List/Reading List/View Controllers/ReadingListTableViewController.swift
@@ -32,7 +32,7 @@ class ReadingListTableViewController: UITableViewController, BookTableViewCellDe
     }
     
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "CellIdentifier", for: indexPath)
+        let cell = tableView.dequeueReusableCell(withIdentifier: "BookCell", for: indexPath)
         guard let bookCell = cell as? BookTableViewCell else { return cell }
         
         let book = bookFor(indexPath: indexPath)
@@ -63,9 +63,9 @@ class ReadingListTableViewController: UITableViewController, BookTableViewCellDe
     
     override func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
         if section == 0 {
-            return "Read Books"
+            return bookController.readBooks.isEmpty ? nil : "Read Books"
         } else {
-            return "Unread Books"
+            return bookController.unreadBooks.isEmpty ? nil : "Unread Books"
         }
     }
